git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/interpreter-3.x/trunk@1076 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ kernel/expression/ExpressionFunction.hpp
@@ -53,7 +53,7 @@
 
 class RexxExpressionFunction : public RexxInternalObject {
  public:
-  RexxExpressionFunction(RexxString *, size_t, RexxQueue *, LONG, BOOL);
+  RexxExpressionFunction(RexxString *, size_t, RexxQueue *, size_t, BOOL);
   inline RexxExpressionFunction(RESTORETYPE restoreType) { ; };
   void        resolve(RexxDirectory *);
   void        live();
